requirements.txt
aiofile==3.9.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
audioop-lts==0.2.2
Authlib==1.6.9
beartype==0.22.9
brotli==1.2.0
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
cyclopts==4.10.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.3
fastmcp==3.2.0
ffmpy==1.0.0
filelock==3.25.2
fsspec==2026.3.0
gradio==6.10.0
gradio_client==2.4.0
groovy==0.1.2
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.13.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
numpy==2.4.4
openai==2.30.0
openapi-pydantic==0.5.1
openenv-core==0.2.3
opentelemetry-api==1.40.0
orjson==3.11.8
packaging==26.0
pandas==3.0.2
pathable==0.5.0
pillow==12.2.0
platformdirs==4.9.4
py-key-value-aio==0.4.4
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.20.0
PyJWT==2.12.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
safehttpx==0.1.7
SecretStorage==3.5.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.4
starlette==0.52.1
tomli==2.4.1
tomli_w==1.2.0
tomlkit==0.13.3
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
zipp==3.23.0
